markup language of gloss words
As discussed in #9 this adds a command-line option to set the language of
the glosses (and potentially other parts of the output besides Greek)
and adds the necessary markup to the output for both the LaTeX and SILE
backends.

Sadly SILE taking generic language codes and LaTeX taking an assortment
of language names (not a normalized set either) it isn't possible to use
the same value for both backends. Rather than expecting users to know
the relevant language names the way LaTeX expects them I suggest using
the standardized short codes and using a lookup table to translate them
into the values needed. This lookup table concept will need to be
expanded as more languages are supported as both typesetting systems
have other options that will need to be loaded for some languages. LaTeX
will need to add `variant=XX` for some language names when it loads the
package and SILE will need to add `script=XX` to the relevant markup to
get the font features right.