switch to KOMA-script class by default in LaTeX backend
While ‘article’ is about as generic as it gets in the world of TeX
document classes, it also brings with it a lot of limitations. The
KOMA-script classes are similar but include a lot more features and
refinements and are less constricting—something newcomers might
appreciated.  The ‘scrartcl’ should be a drop in replacement and be a
little nicer to newcomers and old TeX hands alike. As the KOMA-script
classes are now pretty standard in virtually every LaTeX distribution
I'm not worried about compatibility.

As an example of how this will be easier, try setting the base font
size for a large print edition:

    \documentclass[a4paper,16pt]{{article}}

This just won't work. In fact you'll get an even smaller font that the
12pt you changed it from because it falls back to a default of 10pt on
the unrecognized value. There is a way to make this work but it involves
resetting the base font size yourself and then activating it, two LaTeX
commands most people aren't going to know off the top of their heads. On
the other hand with a KOMA-script class:

    \documentclass[a4paper,16pt]{{scrartcl}}

…everything is going to work as expected.