add support for all languages known by either LaTeX or SILE
Section 2.3 of the [`polyglossia` manual][pg] has a list of supported
languages. Of course this list is maddeningly similar but not identical
to what [`babel`][ba] supports. But as most languages are better supported
in `polyglossia` than in `babel` we're going to ignore that.

SILE supports a [different set of languages][sl]. Somebody was kind
enough to port a bunch over from the LaTeX world, but ① not all of them
came over and ② there are now some that LaTeX doesn't support.
Furthermore SILE uses two letter language codes.

Now you might be thinking, "Oh cool, those must be [ISO 639-1][iso1]."
You'd be wrong. In a [clbuttic][clbuttic] case of [too many cooks in the
kitchen][xkcd] the file set is just alphabet soup¹. There are two letter
codes in there for languages that don't even exist in 639-1 and
differences among those that are. It looks like they might line up with
one of the early RFC's used by W3C, but they aren't a match for the
current [BCP][bcp] either, and even if they were that RFC has issues
and known conflicts.

So neither LaTeX nor SILE currently² supports anything like a standard
list of language names or codes that will translate to the other. No
matter which set was used in this app as canonical I would have to make
up values to map between them.

Instead of filling in blanks I propose using a third—already
complete and normalized—set of codes. For now I have gone with
[ISO-639-3][iso3] codes³ (e.g. `eng`) and mapping to both LaTeX and SILE
values from there. This way we can actually cover all the languages
supported by each platform _and_ document a fixed list that will line up
with _something_ out there.

* ¹ <sub>One must sympathize with Piedmontese here.</sub>
* ² <sub>LaTeX is too far gone, but I plan on fixing this for SILE.
         Until then…</sub>
* ³ <sub>Other options would be to use the hot mess that is the current
         [IETF][ietf] code recommendations (e.g. `en-US`). Franky which RFC we
         go by doesn't matter as long as we pick one that is reasonably future
         proof and that isn't actively being phased out of other systems.</sub>

  [pg]: http://ctan.mirrorcatalogs.com/macros/latex/contrib/polyglossia/polyglossia.pdf
  [ba]: http://people.phys.ethz.ch/~ihn/latex/babel.pdf
  [sl]: https://github.com/simoncozens/sile/tree/master/languages
  [iso1]: https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes
  [xkcd]: https://xkcd.com/927/
  [clbuttic]: https://en.wikipedia.org/wiki/Scunthorpe_problem
  [ietf]: https://en.wikipedia.org/wiki/IETF_language_tag
  [bcp]: http://www.rfc-editor.org/rfc/bcp/bcp47.txt