add whitespace fix to SILE output
Note that whitespace handling in the SILE language is still not quite
nailed down. This extra hack should work on both the last released
version and the current git HEAD version, but there is no promise that
it will be the preferred syntax down the road. However it works for now
and as the backend is still experimental it seems like the smartest
thing to do is use the syntax most likely to work.

See upstream issue 105 for details:

https://github.com/simoncozens/sile/issues/105